Made Amazon S3 uploader more robust, additional error handling. Moved classes to their own files.
